refactor: merge and reorganize mixins
