+++ packages/date-picker/src/vaadin-date-picker.d.ts
@@ -4,10 +4,7 @@
  * This program is available under Apache License Version 2.0, available at https://vaadin.com/license/
  */
 import { ElementMixin } from '@vaadin/component-base/src/element-mixin.js';
-import { ClearButtonMixin } from '@vaadin/field-base/src/clear-button-mixin.js';
-import { FieldAriaMixin } from '@vaadin/field-base/src/field-aria-mixin.js';
-import { InputConstraintsMixin } from '@vaadin/field-base/src/input-constraints-mixin.js';
-import { LabelMixin } from '@vaadin/field-base/src/label-mixin.js';
+import { InputControlMixin } from '@vaadin/field-base/src/input-control-mixin.js';
 import { ThemableMixin } from '@vaadin/vaadin-themable-mixin/vaadin-themable-mixin.js';
 import { DatePickerMixin } from './vaadin-date-picker-mixin.js';
 import { DatePickerEventMap } from './interfaces';
@@ -104,15 +101,7 @@ declare class DatePicker extends HTMLElement {
   ): void;
 }
 
-interface DatePicker
-  extends DatePickerMixin,
-    ElementMixin,
-    ClearButtonMixin,
-    FieldAriaMixin,
-    LabelMixin,
-    DatePickerMixin,
-    InputConstraintsMixin,
-    ThemableMixin {}
+interface DatePicker extends DatePickerMixin, ElementMixin, InputControlMixin, ThemableMixin {}
 
 declare global {
   interface HTMLElementTagNameMap {
