+++ packages/field-base/index.js
@@ -1,18 +1,15 @@
 export { AriaLabelController } from './src/aria-label-controller.js';
 export { CharLengthMixin } from './src/char-length-mixin.js';
 export { CheckedMixin } from './src/checked-mixin.js';
-export { ClearButtonMixin } from './src/clear-button-mixin.js';
 export { DelegateFocusMixin } from './src/delegate-focus-mixin.js';
-export { DelegateInputStateMixin } from './src/delegate-input-state-mixin.js';
 export { DelegateStateMixin } from './src/delegate-state-mixin.js';
-export { FieldAriaMixin } from './src/field-aria-mixin.js';
-export { HelperTextMixin } from './src/helper-text-mixin.js';
+export { FieldMixin } from './src/field-mixin.js';
 export { InputController } from './src/input-controller.js';
+export { InputControlMixin } from './src/input-control-mixin.js';
 export { InputFieldMixin } from './src/input-field-mixin.js';
 export { InputMixin } from './src/input-mixin.js';
 export { LabelMixin } from './src/label-mixin.js';
 export { PatternMixin } from './src/pattern-mixin.js';
 export { SlotStylesMixin } from './src/slot-styles-mixin.js';
 export { TextAreaController } from './src/text-area-controller.js';
-export { TextFieldMixin } from './src/text-field-mixin.js';
 export { ValidateMixin } from './src/validate-mixin.js';
