+++ packages/field-base/src/delegate-state-mixin.js
@@ -15,7 +15,6 @@ const DelegateStateMixinImplementation = (superclass) =>
          */
         stateTarget: {
           type: Object,
-          readOnly: true,
           observer: '_stateTargetChanged'
         }
       };
