+++ packages/field-base/src/input-constraints-mixin.d.ts
@@ -3,7 +3,9 @@
  * Copyright (c) 2021 Vaadin Ltd.
  * This program is available under Apache License Version 2.0, available at https://vaadin.com/license/
  */
-import { DelegateInputStateMixin } from './delegate-input-state-mixin.js';
+import { DelegateStateMixin } from './delegate-state-mixin.js';
+import { InputMixin } from './input-mixin.js';
+import { ValidateMixin } from './validate-mixin.js';
 
 /**
  * A mixin to combine multiple input validation constraints.
@@ -16,7 +18,7 @@ interface InputConstraintsMixinConstructor {
   new (...args: any[]): InputConstraintsMixin;
 }
 
-interface InputConstraintsMixin extends DelegateInputStateMixin {
+interface InputConstraintsMixin extends DelegateStateMixin, InputMixin, ValidateMixin {
   /**
    * Returns true if the current input value satisfies all constraints (if any).
    */
