+++ packages/field-base/src/input-field-mixin.js
@@ -4,18 +4,10 @@
  * This program is available under Apache License Version 2.0, available at https://vaadin.com/license/
  */
 import { dedupingMixin } from '@polymer/polymer/lib/utils/mixin.js';
-import { Debouncer } from '@vaadin/component-base/src/debounce.js';
-import { animationFrame } from '@vaadin/component-base/src/async.js';
-import { ClearButtonMixin } from './clear-button-mixin.js';
-import { DelegateFocusMixin } from './delegate-focus-mixin.js';
-import { FieldAriaMixin } from './field-aria-mixin.js';
-import { LabelMixin } from '../src/label-mixin.js';
-import { InputConstraintsMixin } from './input-constraints-mixin.js';
+import { InputControlMixin } from './input-control-mixin.js';
 
 const InputFieldMixinImplementation = (superclass) =>
-  class InputFieldMixinClass extends ClearButtonMixin(
-    FieldAriaMixin(LabelMixin(InputConstraintsMixin(DelegateFocusMixin(superclass))))
-  ) {
+  class InputFieldMixinClass extends InputControlMixin(superclass) {
     static get properties() {
       return {
         /**
@@ -49,38 +41,12 @@ const InputFieldMixinImplementation = (superclass) =>
          */
         autocapitalize: {
           type: String
-        },
-
-        /**
-         * Specify that the value should be automatically selected when the field gains focus.
-         */
-        autoselect: {
-          type: Boolean,
-          value: false
         }
       };
     }
 
     static get delegateAttrs() {
       return [...super.delegateAttrs, 'autocapitalize', 'autocomplete', 'autocorrect'];
-    }
-
-    static get observers() {
-      return ['__observeOffsetHeight(errorMessage, invalid, label, helperText)'];
-    }
-
-    /** @protected */
-    ready() {
-      super.ready();
-
-      // Lumo theme defines a max-height transition for the "error-message"
-      // part on invalid state change.
-      const errorPart = this.shadowRoot.querySelector('[part="error-message"]');
-      if (errorPart) {
-        errorPart.addEventListener('transitionend', () => {
-          this.__observeOffsetHeight();
-        });
-      }
     }
 
     /**
@@ -119,49 +85,10 @@ const InputFieldMixinImplementation = (superclass) =>
      * @protected
      * @override
      */
-    _onFocus(event) {
-      super._onFocus(event);
-
-      if (this.autoselect && this.inputElement) {
-        this.inputElement.select();
-      }
-    }
-
-    /**
-     * Override an event listener from `DelegateFocusMixin`.
-     * @param {FocusEvent} event
-     * @protected
-     * @override
-     */
     _onBlur(event) {
       super._onBlur(event);
 
       this.validate();
-    }
-
-    /**
-     * Dispatch an event if a specific size measurement property has changed.
-     * Supporting multiple properties here is needed for `vaadin-text-area`.
-     * @protected
-     */
-    _dispatchIronResizeEventIfNeeded(prop, value) {
-      const oldSize = '__old' + prop;
-      if (this[oldSize] !== undefined && this[oldSize] !== value) {
-        this.dispatchEvent(new CustomEvent('iron-resize', { bubbles: true, composed: true }));
-      }
-
-      this[oldSize] = value;
-    }
-
-    /** @private */
-    __observeOffsetHeight() {
-      this.__observeOffsetHeightDebouncer = Debouncer.debounce(
-        this.__observeOffsetHeightDebouncer,
-        animationFrame,
-        () => {
-          this._dispatchIronResizeEventIfNeeded('Height', this.offsetHeight);
-        }
-      );
     }
 
     /**
