+++ packages/field-base/src/input-mixin.d.ts
@@ -17,8 +17,7 @@ interface InputMixinConstructor {
 interface InputMixin {
   /**
    * A reference to the input element controlled by the mixin.
-   * Any component implementing this mixin is expected to provide it
-   * by using `this._setInputElement(input)` Polymer API.
+   * Any component implementing this mixin is expected to provide it.
    *
    * A typical case is using `InputController` that does this automatically.
    * However, the input element does not have to always be native <input>:
