+++ packages/field-base/src/input-mixin.js
@@ -11,8 +11,7 @@ const InputMixinImplementation = (superclass) =>
       return {
         /**
          * A reference to the input element controlled by the mixin.
-         * Any component implementing this mixin is expected to provide it
-         * by using `this._setInputElement(input)` Polymer API.
+         * Any component implementing this mixin is expected to provide it.
          *
          * A typical case is using `InputController` that does this automatically.
          * However, the input element does not have to always be native <input>:
@@ -23,7 +22,6 @@ const InputMixinImplementation = (superclass) =>
          */
         inputElement: {
           type: Object,
-          readOnly: true,
           observer: '_inputElementChanged'
         },
