+++ packages/field-base/src/text-area-controller.js
@@ -25,16 +25,16 @@ export class TextAreaController extends SlotController {
         // rely on `ready()` as we add controller after `ready()` for all mixins has been called.
 
         // InputMixin
-        host._setInputElement && host._setInputElement(node);
+        host.inputElement = node;
 
         // DelegateFocusMixin
-        host._setFocusElement && host._setFocusElement(node);
+        host.focusElement = node;
 
         // DelegateStateMixin
-        host._setStateTarget && host._setStateTarget(node);
+        host.stateTarget = node;
 
         // FieldAriaMixin
-        host._setAriaTarget && host._setAriaTarget(node);
+        host.ariaTarget = node;
       }
     ]);
   }
