+++ packages/field-base/src/validate-mixin.js
@@ -4,10 +4,9 @@
  * This program is available under Apache License Version 2.0, available at https://vaadin.com/license/
  */
 import { dedupingMixin } from '@polymer/polymer/lib/utils/mixin.js';
-import { SlotMixin } from '@vaadin/component-base/src/slot-mixin.js';
 
 const ValidateMixinImplementation = (superclass) =>
-  class ValidateMixinClass extends SlotMixin(superclass) {
+  class ValidateMixinClass extends superclass {
     static get properties() {
       return {
         /**
@@ -26,59 +25,8 @@ const ValidateMixinImplementation = (superclass) =>
         required: {
           type: Boolean,
           reflectToAttribute: true
-        },
-
-        /**
-         * Error to show when the field is invalid.
-         *
-         * @attr {string} error-message
-         */
-        errorMessage: {
-          type: String
         }
       };
-    }
-
-    get slots() {
-      return {
-        ...super.slots,
-        'error-message': () => {
-          const error = document.createElement('div');
-          error.textContent = this.errorMessage;
-          error.setAttribute('aria-live', 'assertive');
-          return error;
-        }
-      };
-    }
-
-    static get observers() {
-      return ['_updateErrorMessage(invalid, errorMessage)'];
-    }
-
-    /** @protected */
-    get _errorNode() {
-      return this._getDirectSlotChild('error-message');
-    }
-
-    constructor() {
-      super();
-
-      // Ensure every instance has unique ID
-      const uniqueId = (ValidateMixinClass._uniqueErrorId = 1 + ValidateMixinClass._uniqueErrorId || 0);
-      this._errorId = `error-${this.localName}-${uniqueId}`;
-    }
-
-    /** @protected */
-    ready() {
-      super.ready();
-
-      if (this._errorNode) {
-        this._errorNode.id = this._errorId;
-
-        this._applyCustomError();
-
-        this._updateErrorMessage(this.invalid, this.errorMessage);
-      }
     }
 
     /**
@@ -97,33 +45,6 @@ const ValidateMixinImplementation = (superclass) =>
      */
     checkValidity() {
       return !this.required || !!this.value;
-    }
-
-    /** @protected */
-    _applyCustomError() {
-      const error = this.__errorMessage;
-      if (error && error !== this.errorMessage) {
-        this.errorMessage = error;
-        delete this.__errorMessage;
-      }
-    }
-
-    /**
-     * @param {boolean} invalid
-     * @protected
-     */
-    _updateErrorMessage(invalid, errorMessage) {
-      if (!this._errorNode) {
-        return;
-      }
-
-      // save the custom error message content
-      if (this._errorNode.textContent && !errorMessage) {
-        this.__errorMessage = this._errorNode.textContent.trim();
-      }
-      const hasError = Boolean(invalid && errorMessage);
-      this._errorNode.textContent = hasError ? errorMessage : '';
-      this.toggleAttribute('has-error-message', hasError);
     }
   };
