+++ packages/field-base/test/field-mixin.test.js
@@ -1,0 +1,479 @@
+import { expect } from '@esm-bundle/chai';
+import { fixtureSync, nextFrame } from '@vaadin/testing-helpers';
+import { PolymerElement, html } from '@polymer/polymer/polymer-element.js';
+import { FieldMixin } from '../src/field-mixin.js';
+import { InputMixin } from '../src/input-mixin.js';
+
+customElements.define(
+  'field-mixin-element',
+  class extends FieldMixin(InputMixin(PolymerElement)) {
+    static get template() {
+      return html`
+        <slot name="label"></slot>
+        <slot name="input"></slot>
+        <slot name="error-message"></slot>
+        <slot name="helper"></slot>
+      `;
+    }
+
+    ready() {
+      super.ready();
+
+      const input = document.createElement('input');
+      input.setAttribute('slot', 'input');
+      this.appendChild(input);
+      this.ariaTarget = input;
+    }
+  }
+);
+
+customElements.define(
+  'field-mixin-group-element',
+  class extends FieldMixin(PolymerElement) {
+    static get template() {
+      return html`
+        <slot name="label"></slot>
+        <slot name="error-message"></slot>
+        <slot name="helper"></slot>
+      `;
+    }
+
+    get _ariaAttr() {
+      return 'aria-labelledby';
+    }
+
+    ready() {
+      super.ready();
+
+      this.ariaTarget = this;
+    }
+  }
+);
+
+describe('field-mixin', () => {
+  let element, label, error, helper, input;
+
+  describe('error message', () => {
+    const ID_REGEX = /^error-field-mixin-element-\d+$/;
+
+    describe('default', () => {
+      beforeEach(() => {
+        element = fixtureSync(`<field-mixin-element></field-mixin-element>`);
+        error = element.querySelector('[slot=error-message]');
+        element.invalid = true;
+      });
+
+      it('should create an error message element', () => {
+        expect(error).to.be.an.instanceof(HTMLDivElement);
+      });
+
+      it('should set aria-live attribute on the error message', () => {
+        expect(error.getAttribute('aria-live')).to.equal('assertive');
+      });
+
+      it('should set id on the error message element', () => {
+        const id = error.getAttribute('id');
+        expect(id).to.match(ID_REGEX);
+        expect(id.endsWith(element.constructor._uniqueFieldId)).to.be.true;
+      });
+
+      it('should update error message content on attribute change', () => {
+        element.setAttribute('error-message', 'This field is required');
+        expect(error.textContent).to.equal('This field is required');
+      });
+
+      it('should update error message content on property change', () => {
+        element.errorMessage = 'This field is required';
+        expect(error.textContent).to.equal('This field is required');
+      });
+
+      it('should clear error message content when field is valid', () => {
+        element.errorMessage = 'This field is required';
+        element.invalid = false;
+        expect(error.textContent).to.equal('');
+      });
+
+      it('should not set has-error-message attribute with no error', () => {
+        expect(element.hasAttribute('has-error-message')).to.be.false;
+      });
+
+      it('should set has-error-message attribute when attribute is set', () => {
+        element.setAttribute('error-message', 'This field is required');
+        expect(element.hasAttribute('has-error-message')).to.be.true;
+      });
+
+      it('should set has-error-message attribute when property is set', () => {
+        element.errorMessage = 'This field is required';
+        expect(element.hasAttribute('has-error-message')).to.be.true;
+      });
+
+      it('should remove has-error-message attribute when field is valid', () => {
+        element.errorMessage = 'This field is required';
+        element.invalid = false;
+        expect(element.hasAttribute('has-error-message')).to.be.false;
+      });
+    });
+
+    describe('attribute', () => {
+      beforeEach(() => {
+        element = fixtureSync(`
+        <field-mixin-element
+          invalid
+          error-message="This field is required"
+        ></field-mixin-element>
+      `);
+        error = element.querySelector('[slot=error-message]');
+      });
+
+      it('should set error-message text content from attribute', () => {
+        expect(error.textContent).to.equal('This field is required');
+      });
+
+      it('should set has-error-message attribute on the host', () => {
+        expect(element.hasAttribute('has-error-message')).to.be.true;
+      });
+    });
+
+    describe('slotted', () => {
+      beforeEach(() => {
+        element = fixtureSync(`
+        <field-mixin-element>
+          <div slot="error-message">Required field</div>
+        </field-mixin-element>
+      `);
+        error = element.querySelector('[slot=error-message]');
+        element.invalid = true;
+      });
+
+      it('should return slotted message content as an errorMessage', () => {
+        expect(element.errorMessage).to.equal('Required field');
+      });
+
+      it('should set id on the slotted error message element', () => {
+        const id = error.getAttribute('id');
+        expect(id).to.match(ID_REGEX);
+        expect(id.endsWith(element.constructor._uniqueFieldId)).to.be.true;
+      });
+
+      it('should set has-error-message attribute with slotted error message', () => {
+        expect(element.hasAttribute('has-error-message')).to.be.true;
+      });
+
+      it('should update slotted error message content on property change', () => {
+        element.errorMessage = 'This field is required';
+        expect(error.textContent).to.equal('This field is required');
+      });
+    });
+  });
+
+  describe('helper', () => {
+    let element, helper;
+
+    const ID_REGEX = /^helper-field-mixin-element-\d+$/;
+
+    describe('default', () => {
+      beforeEach(() => {
+        element = fixtureSync(`<field-mixin-element></field-mixin-element>`);
+        helper = element.querySelector('[slot=helper]');
+      });
+
+      it('should create a helper element', () => {
+        expect(helper).to.be.an.instanceof(HTMLDivElement);
+      });
+
+      it('should set slot on the helper', () => {
+        expect(helper.getAttribute('slot')).to.equal('helper');
+      });
+
+      it('should set id on the helper element', () => {
+        const id = helper.getAttribute('id');
+        expect(id).to.match(ID_REGEX);
+        expect(id.endsWith(element.constructor._uniqueFieldId)).to.be.true;
+      });
+
+      it('should update helper content on attribute change', () => {
+        element.setAttribute('helper-text', '3 digits');
+        expect(helper.textContent).to.equal('3 digits');
+      });
+
+      it('should update helper content on property change', () => {
+        element.helperText = '3 digits';
+        expect(helper.textContent).to.equal('3 digits');
+      });
+
+      it('should not set has-helper attribute with no helper', () => {
+        expect(element.hasAttribute('has-helper')).to.be.false;
+      });
+
+      it('should set has-helper attribute when attribute is set', () => {
+        element.setAttribute('helper-text', '3 digits');
+        expect(element.hasAttribute('has-helper')).to.be.true;
+      });
+
+      it('should set has-helper attribute when property is set', () => {
+        element.helperText = '3 digits';
+        expect(element.hasAttribute('has-helper')).to.be.true;
+      });
+    });
+
+    describe('attribute', () => {
+      beforeEach(() => {
+        element = fixtureSync(`
+          <field-mixin-element
+            helper-text="3 digits"
+          ></field-mixin-element>
+        `);
+        helper = element.querySelector('[slot=helper]');
+      });
+
+      it('should set helper text content from attribute', () => {
+        expect(helper.textContent).to.equal('3 digits');
+      });
+    });
+
+    describe('slotted', () => {
+      beforeEach(() => {
+        element = fixtureSync(`
+          <field-mixin-element>
+            <div slot="helper">Custom</div>
+          </field-mixin-element>
+        `);
+        helper = element.querySelector('[slot=helper]');
+      });
+
+      it('should return slotted helper content as a helperText', () => {
+        expect(element.helperText).to.equal('Custom');
+      });
+
+      it('should set id on the slotted helper element', () => {
+        const id = helper.getAttribute('id');
+        expect(id).to.match(ID_REGEX);
+        expect(id.endsWith(element.constructor._uniqueFieldId)).to.be.true;
+      });
+
+      it('should set has-helper attribute with slotted helper', () => {
+        expect(element.hasAttribute('has-helper')).to.be.true;
+      });
+
+      it('should update slotted helper content on property change', () => {
+        element.helperText = '3 digits';
+        expect(helper.textContent).to.equal('3 digits');
+      });
+    });
+
+    describe('lazy', () => {
+      describe('DOM manipulations', () => {
+        let defaultHelper;
+
+        beforeEach(async () => {
+          element = fixtureSync('<field-mixin-element></field-mixin-element>');
+          await nextFrame();
+          defaultHelper = element._helperNode;
+          helper = document.createElement('div');
+          helper.setAttribute('slot', 'helper');
+          helper.textContent = 'Lazy';
+        });
+
+        it('should return lazy helper content as a helperText using appendChild', async () => {
+          element.appendChild(helper);
+          await nextFrame();
+          expect(element.helperText).to.equal('Lazy');
+        });
+
+        it('should return lazy helper content as a helperText using insertBefore', async () => {
+          element.insertBefore(helper, defaultHelper);
+          await nextFrame();
+          expect(element.helperText).to.equal('Lazy');
+        });
+
+        it('should return lazy helper content as a helperText using replaceChild', async () => {
+          element.replaceChild(helper, defaultHelper);
+          await nextFrame();
+          expect(element.helperText).to.equal('Lazy');
+        });
+
+        it('should remove the default helper from the element when using appendChild', async () => {
+          element.appendChild(helper);
+          await nextFrame();
+          expect(defaultHelper.parentNode).to.be.null;
+        });
+
+        it('should remove the default helper from the element when using insertBefore', async () => {
+          element.insertBefore(helper, defaultHelper);
+          await nextFrame();
+          expect(defaultHelper.parentNode).to.be.null;
+        });
+
+        it('should support replacing lazy helper with a new one using appendChild', async () => {
+          element.appendChild(helper);
+          await nextFrame();
+
+          const div = document.createElement('div');
+          div.setAttribute('slot', 'helper');
+          div.textContent = 'New';
+          element.appendChild(div);
+          await nextFrame();
+          expect(element.helperText).to.equal('New');
+        });
+
+        it('should support replacing lazy helper with a new one using insertBefore', async () => {
+          element.appendChild(helper);
+          await nextFrame();
+
+          const div = document.createElement('div');
+          div.setAttribute('slot', 'helper');
+          div.textContent = 'New';
+          element.insertBefore(div, helper);
+          await nextFrame();
+          expect(element.helperText).to.equal('New');
+        });
+
+        it('should support replacing lazy helper with a new one using replaceChild', async () => {
+          element.appendChild(helper);
+          await nextFrame();
+
+          const div = document.createElement('div');
+          div.setAttribute('slot', 'helper');
+          div.textContent = 'New';
+          element.replaceChild(div, helper);
+          await nextFrame();
+          expect(element.helperText).to.equal('New');
+        });
+      });
+
+      describe('ID attribute', () => {
+        beforeEach(async () => {
+          element = fixtureSync('<field-mixin-element></field-mixin-element>');
+          await nextFrame();
+          helper = document.createElement('div');
+          helper.setAttribute('slot', 'helper');
+          helper.textContent = 'Lazy';
+        });
+
+        it('should set id on the lazily added helper element', async () => {
+          element.appendChild(helper);
+          await nextFrame();
+          expect(helper.getAttribute('id')).to.match(ID_REGEX);
+        });
+
+        it('should not override custom id on the lazily added helper', async () => {
+          helper.id = 'helper-component';
+          element.appendChild(helper);
+          await nextFrame();
+          expect(helper.getAttribute('id')).to.equal('helper-component');
+        });
+
+        it('should restore default id if the custom helper id is removed', async () => {
+          helper.id = 'helper-component';
+          element.appendChild(helper);
+          await nextFrame();
+          helper.removeAttribute('id');
+          await nextFrame();
+          expect(helper.getAttribute('id')).to.match(ID_REGEX);
+        });
+      });
+
+      describe('attributes', () => {
+        beforeEach(async () => {
+          element = fixtureSync('<field-mixin-element></field-mixin-element>');
+          await nextFrame();
+          helper = document.createElement('div');
+          helper.setAttribute('slot', 'helper');
+          helper.textContent = 'Lazy';
+          element.appendChild(helper);
+          await nextFrame();
+        });
+
+        it('should store a reference to the lazily added helper', () => {
+          expect(element._helperNode).to.equal(helper);
+        });
+
+        it('should set has-helper attribute with lazy helper', () => {
+          expect(element.hasAttribute('has-helper')).to.be.true;
+        });
+
+        it('should update lazy helper content on property change', () => {
+          element.helperText = '3 digits';
+          expect(helper.textContent).to.equal('3 digits');
+        });
+      });
+    });
+  });
+
+  describe('aria-describedby', () => {
+    beforeEach(() => {
+      element = fixtureSync(`<field-mixin-element></field-mixin-element>`);
+      label = element.querySelector('[slot=label]');
+      error = element.querySelector('[slot=error-message]');
+      helper = element.querySelector('[slot=helper]');
+      input = element.querySelector('[slot=input]');
+    });
+
+    it('should only add helper text to aria-describedby when field is valid', () => {
+      const aria = input.getAttribute('aria-describedby');
+      expect(aria).to.include(helper.id);
+      expect(aria).to.not.include(error.id);
+      expect(aria).to.not.include(label.id);
+    });
+
+    it('should add error message to aria-describedby when field is invalid', () => {
+      element.invalid = true;
+      const aria = input.getAttribute('aria-describedby');
+      expect(aria).to.include(helper.id);
+      expect(aria).to.include(error.id);
+      expect(aria).to.not.include(label.id);
+    });
+  });
+
+  describe('aria-labelledby', () => {
+    beforeEach(() => {
+      element = fixtureSync(`<field-mixin-group-element></field-mixin-group-element>`);
+      label = element.querySelector('[slot=label]');
+      error = element.querySelector('[slot=error-message]');
+      helper = element.querySelector('[slot=helper]');
+    });
+
+    it('should add label and helper text to aria-labelledby when field is valid', () => {
+      const aria = element.getAttribute('aria-labelledby');
+      expect(aria).to.include(helper.id);
+      expect(aria).to.not.include(error.id);
+      expect(aria).to.include(label.id);
+    });
+
+    it('should add error message to aria-labelledby when field is invalid', () => {
+      element.invalid = true;
+      const aria = element.getAttribute('aria-labelledby');
+      expect(aria).to.include(helper.id);
+      expect(aria).to.include(error.id);
+      expect(aria).to.include(label.id);
+    });
+  });
+
+  describe('custom helper', () => {
+    beforeEach(() => {
+      element = fixtureSync(`<field-mixin-element></field-mixin-element>`);
+      label = element.querySelector('[slot=label]');
+      error = element.querySelector('[slot=error-message]');
+      input = element.querySelector('[slot=input]');
+    });
+
+    it('should handle custom id of a lazily added helper', async () => {
+      helper = document.createElement('div');
+      helper.setAttribute('slot', 'helper');
+      helper.id = 'helper-component';
+      element.appendChild(helper);
+      await nextFrame();
+      expect(input.getAttribute('aria-describedby')).to.include('helper-component');
+    });
+
+    it('should handle restored id of a lazily added helper', async () => {
+      helper = document.createElement('div');
+      helper.setAttribute('slot', 'helper');
+      helper.id = 'helper-component';
+      element.appendChild(helper);
+      await nextFrame();
+      helper.removeAttribute('id');
+      expect(input.getAttribute('aria-describedby')).to.include(helper.id);
+    });
+  });
+});
