+++ packages/text-field/src/vaadin-text-field.d.ts
@@ -4,10 +4,13 @@
  * This program is available under Apache License Version 2.0, available at https://vaadin.com/license/
  */
 import { ElementMixin } from '@vaadin/component-base/src/element-mixin.js';
-import { LabelMixin } from '@vaadin/field-base/src/label-mixin.js';
-import { TextFieldMixin } from '@vaadin/field-base/src/text-field-mixin.js';
+import { CharLengthMixin } from '@vaadin/field-base/src/char-length-mixin.js';
+import { InputFieldMixin } from '@vaadin/field-base/src/input-field-mixin.js';
+import { PatternMixin } from '@vaadin/field-base/src/pattern-mixin.js';
 import { ThemableMixin } from '@vaadin/vaadin-themable-mixin/vaadin-themable-mixin.js';
 
-declare class TextField extends TextFieldMixin(LabelMixin(ThemableMixin(ElementMixin(HTMLElement)))) {}
+declare class TextField extends InputFieldMixin(
+  CharLengthMixin(PatternMixin(ThemableMixin(ElementMixin(HTMLElement))))
+) {}
 
 export { TextField };
